Development

+++ SMS-Backend/src/main/java/com/jamsy/shop/repository/ProductRepository.java
@@ -1,10 +1,10 @@
 package com.jamsy.shop.repository;
 
-import com.jamsy.shop.model.Product;
+import com.jamsy.shop.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    //CRUD
 }
